fix: move SetupDiagnosticsTests to Wizard.Tests assembly, fix accessibility
- CheckPythonForSource: public → internal (Source enum is internal)
- Tests moved from Editor/Tests → Editor/Wizard/Tests (correct asmdef)
- Inline temp dir instead of TempDirScope (cross-assembly dependency)
