--- unity-plugin/Editor/Tests/SetupDiagnosticsTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23f22587b7a1482c8d62a7cdd19b75d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
-  assetBundleName:
-  assetBundleVariant:
